Testen SQL en Integer overal

+++ Tablr/src/sql/MinusExpression.java
@@ -11,29 +11,29 @@ import exceptions.InvalidQueryException;
  * A MinusExpression evaluates to the difference between its left and right subexpression
  *
  */
-public class MinusExpression extends Expression<Double> {
+public class MinusExpression extends Expression<Integer> {
 	/**
 	 * Creates a new MinusExpression.
 	 * @param expression1		Left subexpression
 	 * @param expression2		Right subexpression
 	 */
-	public MinusExpression(Expression<Double> expression1, Expression<Double> expression2) {
+	public MinusExpression(Expression<Integer> expression1, Expression<Integer> expression2) {
 		this.expression1 = expression1;
 		this.expression2 = expression2;
 	}
 	
-	private Expression<Double> expression1;
-	private Expression<Double> expression2;
+	private Expression<Integer> expression1;
+	private Expression<Integer> expression2;
 	
 	/**
 	 * Evalutes this MinusExpression
 	 */
-	public Double eval(ArrayList<Table> tables, int rowNb, HashMap<String,String> tableNames) throws InvalidQueryException {
-		double val1;
-		double val2;
+	public Integer eval(ArrayList<Table> tables, int rowNb, HashMap<String,String> tableNames) throws InvalidQueryException {
+		Integer val1;
+		Integer val2;
 		try {
-			val1 = (Double)expression1.eval(tables, rowNb, tableNames);
-			val2 = (Double)expression2.eval(tables, rowNb, tableNames);
+			val1 = (Integer)expression1.eval(tables, rowNb, tableNames);
+			val2 = (Integer)expression2.eval(tables, rowNb, tableNames);
 		} catch (ClassCastException e) {
 			throw new InvalidQueryException("Can't subtract values in MinusExpression: invalid types");
 		}
